requirements.txt

streamlit
PyJWT
